Moved classes to their rightful packages plus a bit of cleanup
