+++ realm/realm-library/src/objectServer/java/io/realm/mongodb/log/obfuscator/EmailPasswordObfuscator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.realm.internal.obfuscator;
+package io.realm.mongodb.log.obfuscator;
 
 import java.util.HashMap;
 import java.util.Map;
