+++ realm/realm-library/src/testObjectServer/kotlin/io/realm/mongodb/log/obfuscator/CustomFunctionObfuscatorTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.realm.internal.obfuscator
+package io.realm.mongodb.log.obfuscator
 
 import org.junit.Test
 import kotlin.test.assertEquals
